Async Task is a Generic class
Three types of parameter in an Async Task are the following:

Params: Parameter type sent to the task upon execution
Progress: Type published to update progress during the background computation
Result: The type pf the result of the background computation

The functions one can overwrite in the Async Task are the following:

1. onPreExecute: Initialized UI thread
2. do InBackground: The long task which you want to happen in the background -> publish progress
3. onProgressUpdate:
4. onPostExecute:

so in this program, first a new object of GithubQueryTask is introduced which extends the AsyncTask. For the AsyncTask the three types of Params, progress and result have to be indicated. In this class they decided to be URL, void and String. It means that the input is a URL and the Output will be an string. The input comes with ... which shows that it is an array.
First the doInbackground method must be overwrited to make the url ready and start the connection. In the postExecute method the message will be displayed on screen. as the whole GithubQueryTask is a class, we need to make a new object with its type,      new GithubQueryTask()
and usign         .execute()
we execute it, keep in mind that the code is written in a way that the execution needs the initial parameters which have the type of URL and that's what we feed into the network.