Adapter for recycleView
To pass the information to the recycle view and adaptor is needed, to make it work three functions must be overwrite:
onCreateViewHolder: called when the RecyclerView make a new viewHolder instance
onBindViewHolder: when the recycle view puts data form our model
getItemCount: returns the number of items in the data source.

when I wanted to over write the functions in the recyclerView, I had to overwrite all three of these functions.
To overwrite the onCreateViewHolder, keep in mind that the return value is a viewHolder as well and onCreateViewHolder must have two inputs, the viewGroup and an int to show the view type.
first we need to define the context that the code is running in using
Context context= parent.getContext()
then we need to get the layout, consider that the position of a layout is a int.
then we need to make an inflator and use it to inflate the layout.
 
The general gist is this: If shouldAttachToParrentImmediately
 is set to true, then the layout file specified in the first parameter is inflated and attached to the ViewGroup specified in the second parameter.
Then the method returns this combined View, with the ViewGroup as the root. When shouldAttachToParrentImmediately
 is false, the layout file from the first parameter is inflated and returned as a View. The root of the returned View would simply be the root specified in the layout file. In either case, the ViewGroup’s LayoutParams are needed to correctly size and position the View created from the layout file.

then a view object generated and inflated with the inflater and according to the last parameter it will not attach to the viewGroup immediately.
then the viewHolder is generated using the constructor which we had defined before and the viewHolder will be set and returns as the output of onCreateViewHolder.