Steps to make a recycle view and some notes to make it more clear (1) Add RecyclerView dependency implementation 'com.android.support:recyclerview-v7:28.0.0' (2) make a RecyclerView in the Layout it is done by RC on layout and make a new layout resource file and then making the root either vertical or horizontal Linear layout. (3) Making the changes in the layout
(4) Making a class for the adapter,
it literally means that make a new file and make a new class in it. RC on utilities in this project -> new -> java class
(5) inside the adaptor class make another class which extends on the RecyclerView.ViewHolder

It’s now a new file and you have to make the class inside it to declare the features. Also the Forecast class is extended over the RecyclerView but it still needs the viewHolders so the new class which we generate inside it is responsible for the ViewHolder.
(6) connect he parents of the view holder (ForecaseAdapterViewHolder) to the textView
it is done using findView
mWeatherTestView= (TextView) itemView.findViewById(R.id.tv_weather_data);
The thing is to remember that the textview in the layout which was generated is connected to the adapter.

(7) build the onCreateViewHolder

public class ForecastAdapter extends RecyclerView.Adapter<ForecastAdapter.ForecastAdapterViewHolder> {

\\ the type of input of the ForecastAdaptor is ForecaseAdapterViewHolder instead of a regular VH.

class ForecastAdapterViewHolder extends  RecyclerView.ViewHolder{

\\ The ForecastAdapterViewHolder has to be declared thoroughly

    public final TextView mWeatherTestView;

\\ this is where we connect the adapter to the TextView so we connect it to the textView defined in the layout designated for this

    public ForecastAdapterViewHolder(@NonNull View itemView) {
        super(itemView);
        mWeatherTestView= (TextView) itemView.findViewById(R.id.tv_weather_data);
    }

\\ it needs a constructor to initialize the value, the constructor will accept the view, and initialize the textView.

}

\\ the next step is to over write the three main functions in an adaptor

@NonNull
@Override
public ForecastAdapterViewHolder onCreateViewHolder(@NonNull ViewGroup viewGroup, int i) {
\\The first thing is to define a context.
    Context context = viewGroup.getContext();
\\ the integer for layoutIdForListItem decides how many item will be in the list based on the layout.
    int layoutIdForListItem= R.layout.forecast_list_item;
\\ the layout has to be inflated, and to do that a layout inflater is needed, it will put the view based on the number assigned to it.
    LayoutInflater inflater =LayoutInflater.from(context);
    boolean shouldAttachToParentImmediately= false;
    View view = inflater.inflate(layoutIdForListItem,viewGroup,
            shouldAttachToParentImmediately);

\\forecastViewHolder is from the nature of the viewHolders so it needs to know which view.
    ForecastAdapterViewHolder  viewHolder = new ForecastAdapterViewHolder(view);
    return viewHolder;
}