Intends The next step is to pass information from the main activity to the child activity, To do that first in mainActivity we need to define the intent which points to the child activity Intent startChildActivityIntent = new Intent(context, destinationActivity);
then we need to pass the information to it.
startChildActivityIntent.putExtra("entryText",textEntery);

and run the intend
startActivity(startChildActivityIntent);

in the child activity, the intent needs to be retried
Intent childActivityIntent = getIntent();

and the information has to be used in the proper locations.
Bundle extras = childActivityIntent.getExtras();
if (childActivityIntent.hasExtra("entryText")) {
    String textFromIntent = extras.getString("entryText");
    mDisplayText.setText(textFromIntent);

}