scheme:[//[user:password@]host[:port}}[/]path[?query][#fragment]
To build a Uril, use a uri builder. Given that the following is the address,

String address= "200 s bonnie brae st. Denton Texas 76201";

first you need to build a uri builder

Uri.Builder builder = new Uri.Builder();

The next step is to provide the appropriate scheme and path, because we don’t have the latitude and longitude we put zero in there.
builder.scheme("geo").path("0,0").query(address);

And then we ask the builder to build it.
Uri addressUri= builder.build();

For some reason it doesn't show the address but this is a problem for another time.